requirements.txt
git+https://github.com/waggle-sensor/pywaggle@v0.39.1


keras
Pillow
numpy
tensorflow
matplotlib

